move utils to common folder; add testutils.py
