+++ dojotTester/tests/example.py
@@ -1,5 +1,5 @@
-from tests.base_test import BaseTest
-from dojot.api import DojotAPI as Api
+from common.testutils import *
+from common.base_test import BaseTest
 from mqtt.mqttClient import MQTTClient
 
 
@@ -24,6 +24,18 @@ class DummyTest(BaseTest):
         super().tearDown()
 
 
+class RegistryASimpleDevice(BaseTest):
+    """
+    API Registry a Simple Device.
+    """
+
+    def runTest(self):
+
+        device_id = add_a_simple_device(self)
+        self.logger.info("device id is: " + device_id)
+        self.assertTrue(device_id is not None, "** FAILED ASSERTION: device id is None")
+
+
 class ApiDummyTest(BaseTest):
     """
     API dummy example.
@@ -33,7 +45,7 @@ class ApiDummyTest(BaseTest):
         self.logger.info('Executing Api dummy test...')
         jwt = Api.get_jwt()
         self.logger.info("JWT = " + jwt)
-        self.assertTrue(jwt is not None, "JWT nulo")
+        self.assertTrue(jwt is not None, "JWT is None")
 
 
 class ApiErrorDummyTest(BaseTest):
@@ -42,12 +54,11 @@ class ApiErrorDummyTest(BaseTest):
     """
 
     def runTest(self):
-        self.logger.info('Executing Api dummy test...')
+
         jwt = Api.get_jwt()
         self.logger.info("JWT = " + jwt)
-        self.assertTrue(jwt is not None, "JWT nulo")
+        self.assertTrue(jwt is not None, "** FAILED ASSERTION: failure to get JWT **")
 
-        # template payload malformed
         template1 = {
             "labe": "SensorModel",
             "attrs": [
@@ -67,7 +78,7 @@ class ApiErrorDummyTest(BaseTest):
         rc, response = Api.create_template(jwt, template1)
         self.logger.info(f"result code is {rc}")
         self.logger.info(f"response is {response}")
-        self.assertTrue(rc == 400, "received a unexpected result code")
+        self.assertTrue(rc == 400, "** FAILED ASSERTION: received an unexpected result code **")
 
 
 class MqttDummyTest(BaseTest):
